+++ tests/test_errors.py
@@ -26,7 +26,19 @@ def test_validate_primer_2():
         oprimer = Primer("ATCGATCGATCGATCGAF")
 
 
-def test_melting_temperature():
+def test_melting_temperature_1():
     oprimer = Primer("ATCGATCGATCGATCGAT")
     with pytest.raises(NotImplementedError, match=r"This method for calculating melting temperature has not been implemented."):
         oprimer.melting_temperature(method=MeltingTemperature.NEAREST_NEIGHBOR)
+
+
+def test_melting_temperature_2():
+    oprimer = Primer("ATCGATCGATCGATCGAT")
+    with pytest.raises(NotImplementedError, match=r"This method for calculating melting temperature has not been implemented."):
+        oprimer.melting_temperature(method="Invalid Method")
+
+
+def test_is_computed():
+    oprimer = Primer("ATCGATCGATCGATCGAT")
+    with pytest.raises(OPRBaseError, match=r"^This attribute either doesn't exist or cannot be computed/cached."):
+        oprimer.is_computed("Invalid Attribute")